[de] Add link to german podcast about dcvs
Link to the german podcast series Chaosradio Express and one of its podcasts
about distributed version control systems. The talk is about dvcs in general
and references Git a lot. It's an exciting peace of work and pretty useful to
get an overview about the topic at all. My recommendation.